+++ rust/processor/src/db/common/models/emojicoin_models/models/market_24h_rolling_volume.rs
@@ -9,8 +9,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct RecentOneMinutePeriodicStateEvent {
-    pub market_id: i64,
-    pub market_nonce: i64,
+    pub market_id: BigDecimal,
+    pub market_nonce: BigDecimal,
     pub transaction_version: i64,
     pub period_quote_volume: BigDecimal,
     pub period_base_volume: BigDecimal,
